requirements.txt
numpy>=1.21.0
pandas>=1.3.0
torch>=2.0.0
scikit-learn>=1.0.0
tqdm>=4.62.0
pyyaml>=6.0
matplotlib>=3.4.0

